Separated server and app content, still works
